# docs/baseline/openapi-backend-v0.2.json
{
  "components": {
    "schemas": {
...
      "AggregateMetrics": {
        "properties": {
          "average_unit_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Unit Price"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "median_unit_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Unit Price"
          },
          "total_amount": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Amount"
          }
        },
        "title": "AggregateMetrics",
        "type": "object"
      },
      "AnalyticsAggregation": {
        "enum": [
          "COUNT",
          "AVERAGE_UNIT_PRICE",
          "MEDIAN_UNIT_PRICE",
          "TOTAL_AMOUNT"
        ],
        "title": "AnalyticsAggregation",
        "type": "string"
      },
      "AnalyticsGroupBy": {
        "enum": [
          "BRAND",
          "BUILDING",
          "SUPPLIER",
          "DEVICE_NAME"
        ],
        "title": "AnalyticsGroupBy",
        "type": "string"
      },
      "AnalyticsSortBy": {
        "enum": [
          "CREATED_AT",
          "UNIT_PRICE",
          "TOTAL_AMOUNT",
          "QUANTITY"
        ],
        "title": "AnalyticsSortBy",
        "type": "string"
      },
...
      "ApiResponse_PurchaseQueryData_": {
        "properties": {
          "code": {
            "default": "OK",
            "title": "Code",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/PurchaseQueryData"
          },
          "message": {
            "default": "操作成功",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[PurchaseQueryData]",
        "type": "object"
      },
...
      "ApiResponse_RequirementRiskData_": {
        "properties": {
          "code": {
            "default": "OK",
            "title": "Code",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/RequirementRiskData"
          },
          "message": {
            "default": "操作成功",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[RequirementRiskData]",
        "type": "object"
      },
...
      "ApiResponse_SimilarCasesData_": {
        "properties": {
          "code": {
            "default": "OK",
            "title": "Code",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/SimilarCasesData"
          },
          "message": {
            "default": "操作成功",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[SimilarCasesData]",
        "type": "object"
      },
...
      "ApiResponse_SupplierPerformanceData_": {
        "properties": {
          "code": {
            "default": "OK",
            "title": "Code",
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/SupplierPerformanceData"
          },
          "message": {
            "default": "操作成功",
            "title": "Message",
            "type": "string"
          },
          "success": {
            "default": true,
            "title": "Success",
            "type": "boolean"
          },
          "trace_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Id"
          }
        },
        "required": [
          "data"
        ],
        "title": "ApiResponse[SupplierPerformanceData]",
        "type": "object"
      },
...
      "GroupedMetrics": {
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "metrics": {
            "$ref": "#/components/schemas/AggregateMetrics"
          }
        },
        "required": [
          "key",
          "label",
          "metrics"
        ],
        "title": "GroupedMetrics",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JsonValue": {},
...
      "PurchaseAnalysisItem": {
        "properties": {
          "actual_total_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Total Price"
          },
          "actual_unit_price": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actual Unit Price"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "building_id": {
            "title": "Building Id",
            "type": "integer"
          },
          "building_name": {
            "title": "Building Name",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "device_profession": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Profession"
          },
          "expected_arrival_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Arrival Date"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "purchased_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchased At"
          },
          "quantity": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "received_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "received_quantity": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received Quantity"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "integer"
          },
          "requirement_no": {
            "title": "Requirement No",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "supplier_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Name"
          }
        },
        "required": [
          "requirement_id",
          "requirement_no",
          "building_id",
          "building_name",
          "device_profession",
          "device_name",
          "brand",
          "model",
          "quantity",
          "status",
          "supplier_id",
          "supplier_name",
          "actual_unit_price",
          "actual_total_price",
          "expected_arrival_date",
          "purchased_at",
          "received_quantity",
          "received_at",
          "created_at",
          "completed_at"
        ],
        "title": "PurchaseAnalysisItem",
        "type": "object"
      },
...
      "PurchaseQueryData": {
        "properties": {
          "effective_query": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Effective Query",
            "type": "object"
          },
          "groups": {
            "items": {
              "$ref": "#/components/schemas/GroupedMetrics"
            },
            "title": "Groups",
            "type": "array"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PurchaseAnalysisItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "scanned_rows": {
            "title": "Scanned Rows",
            "type": "integer"
          },
          "summary": {
            "$ref": "#/components/schemas/AggregateMetrics"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "items",
          "summary",
          "groups",
          "page",
          "page_size",
          "total",
          "scanned_rows",
          "effective_query"
        ],
        "title": "PurchaseQueryData",
        "type": "object"
      },
      "PurchaseQueryRequest": {
        "additionalProperties": false,
        "properties": {
          "aggregations": {
            "items": {
              "$ref": "#/components/schemas/AnalyticsAggregation"
            },
            "maxItems": 4,
            "minItems": 1,
            "title": "Aggregations",
            "type": "array"
          },
          "brands": {
            "items": {
              "maxLength": 100,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "title": "Brands",
            "type": "array"
          },
          "building_ids": {
            "items": {
              "exclusiveMinimum": 0.0,
              "type": "integer"
            },
            "maxItems": 50,
            "title": "Building Ids",
            "type": "array"
          },
          "created_from": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created From"
          },
          "created_to": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created To"
          },
          "device_name": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Name"
          },
          "device_professions": {
            "items": {
              "enum": [
                "电气",
                "暖通",
                "弱电",
                "机房环境",
                "工器具",
                "算力服务器",
                "IDC网络",
                "其他"
              ],
              "type": "string"
            },
            "maxItems": 20,
            "title": "Device Professions",
            "type": "array"
          },
          "exclude_blacklisted": {
            "default": false,
            "title": "Exclude Blacklisted",
            "type": "boolean"
          },
          "exclude_delayed_suppliers": {
            "default": false,
            "title": "Exclude Delayed Suppliers",
            "type": "boolean"
          },
          "group_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalyticsGroupBy"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_total_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Total Price"
          },
          "max_unit_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Unit Price"
          },
          "min_total_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Total Price"
          },
          "min_unit_price": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Unit Price"
          },
          "models": {
            "items": {
              "maxLength": 150,
              "minLength": 1,
              "type": "string"
            },
            "maxItems": 50,
            "title": "Models",
            "type": "array"
          },
          "page": {
            "default": 1,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "default": 20,
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Page Size",
            "type": "integer"
          },
          "sort_by": {
            "$ref": "#/components/schemas/AnalyticsSortBy",
            "default": "CREATED_AT"
          },
          "sort_order": {
            "$ref": "#/components/schemas/SortOrder",
            "default": "DESC"
          },
          "statuses": {
            "items": {
              "$ref": "#/components/schemas/PurchaseStatus"
            },
            "maxItems": 20,
            "title": "Statuses",
            "type": "array"
          },
          "supplier_ids": {
            "items": {
              "exclusiveMinimum": 0.0,
              "type": "integer"
            },
            "maxItems": 50,
            "title": "Supplier Ids",
            "type": "array"
          }
        },
        "title": "PurchaseQueryRequest",
        "type": "object"
      },
...
      "PurchaseStatus": {
        "enum": [
          "DRAFT",
          "PENDING_REVIEW",
          "REJECTED",
          "PENDING_PURCHASE",
          "PURCHASING",
          "PENDING_WAREHOUSE",
          "COMPLETED"
        ],
        "title": "PurchaseStatus",
        "type": "string"
      },
      "RatioMetric": {
        "properties": {
          "denominator": {
            "title": "Denominator",
            "type": "integer"
          },
          "numerator": {
            "title": "Numerator",
            "type": "integer"
          },
          "ratio": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio"
          }
        },
        "required": [
          "numerator",
          "denominator",
          "ratio"
        ],
        "title": "RatioMetric",
        "type": "object"
      },
...
      "RequirementRiskData": {
        "properties": {
          "evaluated_at": {
            "format": "date-time",
            "title": "Evaluated At",
            "type": "string"
          },
          "matched_count": {
            "title": "Matched Count",
            "type": "integer"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "integer"
          },
          "scanned_rows": {
            "title": "Scanned Rows",
            "type": "integer"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/RiskSignal"
            },
            "title": "Signals",
            "type": "array"
          }
        },
        "required": [
          "requirement_id",
          "evaluated_at",
          "signals",
          "matched_count",
          "scanned_rows"
        ],
        "title": "RequirementRiskData",
        "type": "object"
      },
...
      "RiskLevel": {
        "enum": [
          "INFO",
          "LOW",
          "MEDIUM",
          "HIGH"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
      "RiskSignal": {
        "properties": {
          "facts": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Facts",
            "type": "object"
          },
          "matched": {
            "title": "Matched",
            "type": "boolean"
          },
          "metrics": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Metrics",
            "type": "object"
          },
          "related_record_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Related Record Ids",
            "type": "array"
          },
          "risk_code": {
            "title": "Risk Code",
            "type": "string"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "risk_type": {
            "title": "Risk Type",
            "type": "string"
          },
          "threshold": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Threshold",
            "type": "object"
          },
          "time_range": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue"
            },
            "title": "Time Range",
            "type": "object"
          }
        },
        "required": [
          "risk_code",
          "risk_type",
          "risk_level",
          "matched",
          "facts",
          "metrics",
          "related_record_ids",
          "threshold",
          "time_range"
        ],
        "title": "RiskSignal",
        "type": "object"
      },
...
      "SimilarCasesData": {
        "properties": {
          "algorithm": {
            "title": "Algorithm",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SimilarCaseItem"
            },
            "title": "Items",
            "type": "array"
          },
          "requirement_id": {
            "title": "Requirement Id",
            "type": "integer"
          },
          "scanned_rows": {
            "title": "Scanned Rows",
            "type": "integer"
          }
        },
        "required": [
          "requirement_id",
          "algorithm",
          "items",
          "scanned_rows"
        ],
        "title": "SimilarCasesData",
        "type": "object"
      },
...
      "SortOrder": {
        "enum": [
          "ASC",
          "DESC"
        ],
        "title": "SortOrder",
        "type": "string"
      },
...
      "SupplierPerformanceData": {
        "properties": {
          "average_delivery_days": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Delivery Days"
          },
          "blacklist_history_count": {
            "title": "Blacklist History Count",
            "type": "integer"
          },
          "building_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Building Ids",
            "type": "array"
          },
          "building_names": {
            "items": {
              "type": "string"
            },
            "title": "Building Names",
            "type": "array"
          },
          "created_from": {
            "format": "date",
            "title": "Created From",
            "type": "string"
          },
          "created_to": {
            "format": "date",
            "title": "Created To",
            "type": "string"
          },
          "current_blacklist_status": {
            "title": "Current Blacklist Status",
            "type": "string"
          },
          "delay": {
            "$ref": "#/components/schemas/RatioMetric"
          },
          "historical_purchase_count": {
            "title": "Historical Purchase Count",
            "type": "integer"
          },
          "last_cooperation_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Cooperation At"
          },
          "quantity_anomaly": {
            "$ref": "#/components/schemas/RatioMetric"
          },
          "supplier_id": {
            "title": "Supplier Id",
            "type": "integer"
          },
          "supplier_name": {
            "title": "Supplier Name",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "supplier_id",
          "supplier_name",
          "created_from",
          "created_to",
          "historical_purchase_count",
          "last_cooperation_at",
          "average_delivery_days",
          "delay",
          "quantity_anomaly",
          "current_blacklist_status",
          "blacklist_history_count",
          "building_ids",
          "building_names"
        ],
        "title": "SupplierPerformanceData",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Procurement Agent",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/analytics/purchase-query": {
      "post": {
        "operationId": "purchase_query_api_v1_analytics_purchase_query_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PurchaseQueryData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Purchase Query",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/v1/requirements/{requirement_id}/risk-signals": {
      "get": {
        "operationId": "requirement_risk_signals_api_v1_requirements__requirement_id__risk_signals_get",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "title": "Requirement Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_RequirementRiskData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Requirement Risk Signals",
        "tags": [
          "analytics"
        ]
      }
    },
    "/api/v1/requirements/{requirement_id}/similar-cases": {
      "get": {
        "operationId": "requirement_similar_cases_api_v1_requirements__requirement_id__similar_cases_get",
        "parameters": [
          {
            "in": "path",
            "name": "requirement_id",
            "required": true,
            "schema": {
              "title": "Requirement Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 20,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SimilarCasesData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Requirement Similar Cases",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/v1/suppliers/{supplier_id}/performance": {
      "get": {
        "operationId": "supplier_performance_api_v1_suppliers__supplier_id__performance_get",
        "parameters": [
          {
            "in": "path",
            "name": "supplier_id",
            "required": true,
            "schema": {
              "title": "Supplier Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "created_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created From"
            }
          },
          {
            "in": "query",
            "name": "created_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_SupplierPerformanceData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Supplier Performance",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}